Fix the use of node labels to determine the cluster name on the DCA (#14232)
