+++ pkg/util/kubernetes/hostinfo/cluster_name_from_node_label_nothing.go
@@ -1,0 +1,18 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+//go:build !kubelet && !kubeapiserver
+// +build !kubelet,!kubeapiserver
+
+package hostinfo
+
+import (
+	"context"
+)
+
+// getANodeLabels returns nothing
+func (_ *NodeInfo) getANodeLabels(_ context.Context) (map[string]string, error) {
+	return nil, nil
+}
